#24 생일 와이어프레임 화면 리팩토링 진행
BirthdayViewController-layout.swift 파일을 생성하여 UI 관련 로직 이동.
